requirements.txt
psycopg2-binary==2.9.7
fastapi==0.104.1
uvicorn==0.24.0